Add initial structure for MailService application.
This commit introduces a large set of files that act as the initial structure for the MailService project. Significant changes include the configuration of mail sender, the handling of mail-related actions, application properties, and mail body templates. Furthermore, a Dockerfile and a docker-compose script are added to set up containerized deployment. Sample testing files are also included.